fix(riscv64): a truncated register hint, and two encoders that lied about their size
spectest v2 now passes in full on riscv64.

Three defects, two of them outside this backend:

  - newDesiredLocReg packs a RealReg two bits up inside a uint16, but
    RealReg is a byte -- so the shift wrapped before the widening, and any
    register above 63 came back as a different one. riscv64's third
    register file starts at 64: v1 (RealReg 65) read back as ra (RealReg
    1), and a v128 phi was reconciled by moving a vector register into an
    integer one. amd64 and arm64 have no register that high and cannot
    reach it.

  - v128.const declared sixteen bytes and emitted twenty-four. size() is
    not a hint -- resolveRelativeAddresses lays out every label from it --
    so an `if` whose arms held v128 constants branched four bytes short and
    ran the wrong arm. The two instructions over budget were a copy to the
    scratch and straight back again, which did nothing; removing them makes
    the encoder match the declaration rather than the other way round.

  - i8x16.bitmask had the same mismatch, off by the two shifts that trim
    the mask's tail.

The last two are a class, not a pair, so encode now checks each
instruction's emitted length against size() and panics on a mismatch. It
caught the bitmask one on the first run. instruction.String() also grew the
sixteen vector cases it was missing, without which the debug printers
panic before printing anything.

+++ internal/engine/native/backend/regalloc/regalloc.go
@@ -176,8 +176,15 @@ const (
 	desiredLocStack       = desiredLoc(desiredLocKindStack)
 )
 
+// newDesiredLocReg packs a register into the two bits above the kind.
+//
+// The widening is not cosmetic: RealReg is a byte, so shifting one that names a
+// register above 63 -- which riscv64's third register file does -- wraps inside
+// the byte before it ever reaches the uint16. RealReg 65 (v1) came back out as
+// RealReg 1 (ra), and a v128 phi was reconciled by moving a vector register
+// into an integer one.
 func newDesiredLocReg(r RealReg) desiredLoc {
-	return desiredLoc(desiredLocKindReg) | desiredLoc(r<<2)
+	return desiredLoc(desiredLocKindReg) | desiredLoc(uint16(r)<<2)
 }
 
 func (d desiredLoc) realReg() RealReg {
